Adds support for manual action file URL
Adds the ability to specify a URL to an XLS file
containing manual actions to be performed during
deployment. This allows including documentation and
instructions directly within the pipeline configuration.

A documentation link has also been added to the UI to allow
users to directly access documentation for each configurable
parameter in the config.