Fixes Code Builder integration & cache manager issues (#226)
* Fixes Code Builder integration & cache manager issues

Addresses issues related to Code Builder integration and cache management.
Specifically, it adjusts the SFDX path check to correctly identify the
Code Builder environment and ensures that cache resets are handled
asynchronously to prevent blocking operations. This also enables the
cache to properly clear, addressing a previous issue.

* [Mega-Linter] Apply linters fixes

* fix org cache

* Improves web VS Code file handling

Detects if the extension is running in web VS Code.

If running in web VS Code, triggers a download for binary files
instead of attempting to open them externally, as opening
externally is not supported in the web environment.

* More logs in output panel

* [Mega-Linter] Apply linters fixes

* Updates dependencies and improves error logging

Updates the minimum recommended version of a dependency to address potential compatibility issues.

Improves error logging for pull request button information by stringifying the error object to provide more detailed diagnostic information. This aids in debugging issues related to retrieving pull request information.

* Remove file download handling and implement download message handling in LwcUiPanel and CommandExecution

* changelog

* Synchronizes panel initialization with user input

Ensures the user input type is sent to the client after the command execution panel is fully initialized.

This change fixes a timing issue where the client might not receive the user input type if it connects before the panel is ready.
It also removes unnecessary null check.

* [Mega-Linter] Apply linters fixes

* Ensures LWC panel is ready before command execution

This change ensures that the LWC panel is fully initialized and ready to receive messages before command execution proceeds. It prevents potential issues caused by sending messages to a panel that is not yet ready.

A 'commandLWCReady' message is sent from the LWC panel to the server when it's ready. The server then sends a 'userInput' message back to the caller to signal that it can continue with the command.

* [Mega-Linter] Apply linters fixes

* changelog

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>