Adds filter to combobox for easier selection (#234)
Improves the user experience for comboboxes with many options by introducing a filter input.
This simplifies the selection process by allowing users to quickly narrow down the list of choices.
The filter input is toggled by a button next to the combobox.

Includes CSS adjustments for proper alignment and sizing of the filter components.