Enhances org management and command execution (#238)
* (todo: highlight default) + Refreshes UI elements on configuration changes

Ensures UI elements reflecting the org status are refreshed when sfdx configuration files are changed, or when the org configuration is updated by the user.

This change improves user experience by automatically updating the org manager and status bar when changes occur to the sfdx configuration.

* Allows overriding duplicate command check

Adds a mechanism to bypass the duplicate command execution check.

This allows users to override the check and force the execution of a command, even if it's already running. This is achieved by introducing a temporary flag that is set when the user explicitly requests to run the command again.

This addresses scenarios where the user believes the previous command execution was faulty or incomplete, and they need to retry it immediately.

* changelog

* [Mega-Linter] Apply linters fixes

* Handle default org

* changelog

* Adjusts column widths in the org table

Adjusts the column widths for "Type" and "Role" in the org table to improve readability and layout.

Reduces the "Type" column width from 120 to 100.
Increases the "Role" column width from 80 to 100.

* [Mega-Linter] Apply linters fixes

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>