Adds unsecured connected apps diagnosis command (#242)
* Adds unsecured connected apps diagnosis command

Adds a new command to diagnose unsecured connected apps.

This change also addresses an issue where errors occurring at the start of a
command execution were not being displayed in the LWC panel.
It now ensures that errors are properly captured and displayed to the user.

It also sends background command ended messages to all panels to update status
and handle errors.

* [Mega-Linter] Apply linters fixes :)

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>