Adds feature branch display toggle
Allows users to toggle the display of feature branches in the DevOps Pipeline diagram.

This change introduces a new configuration option that controls whether feature branches are displayed. It enhances the user experience by allowing users to focus on major branches or view all branches depending on their needs.

It introduces a new VS Code setting `vsCodeSfdxHardis.pipelineDisplayFeatureBranches`
to control this behavior.

Adds emojis to the PR labels to indicate the status of the linked Jobs