Adds autoclose command feature (#272)
* Adds autoclose command feature

Implements a new feature that allows users to configure commands to automatically close their execution panel upon successful completion. This improves the user experience by reducing screen clutter and streamlining the workflow.

The feature introduces a new configuration setting, "autocloseCommands", which is a list of commands that should be automatically closed. A toggle has been added to the command execution panel, enabling users to add or remove the current command from the autoclose list directly from the UI.

* [Mega-Linter] Apply linters fixes :)

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>