View merged PRs in mermaid pipeline
Fetches pull requests merged into each branch since the last merge to target branch.

This enhancement provides visibility into the MRs included in each environment,
listing pull requests merged in the branch itself and child branches up to the
last merge with the next environment in the pipeline.

Includes a method to recursively retrieves child branches names.

Adds an indicator in the mermaid graph to show the number of MRs per environment.