Enables lazy-loading of metadata types
Introduces lazy-loading of metadata types in the metadata retriever to improve performance. It also adds `IsNameObsolete` to the SourceMember query to correctly identify deleted members.

The metadata types are now fetched when the org is selected, and when the org is changed, which reduces initial loading time and ensures that the list of available metadata types is up-to-date for the selected org.