Improve Metadata Retriever (#276)
* Improves metadata retriever functionality

Enhances the metadata retriever to display the operation performed on metadata members (created, modified, deleted).

Adds an icon to indicate the operation type in "Recent Changes" mode.

Shows a message when no search has been performed yet.

Clears results and resets search state when the org or mode is changed.

* Adds local file check to metadata retriever

Enables checking for local files in the metadata retriever UI.

Adds a toggle to the metadata retriever panel that enables users to check if retrieved metadata files exist locally. The feature checks the sfdx-project.json for package directories and annotates the results with a "LocalFileExists" flag. The UI displays a checkmark icon to indicate the existence of a local file. Improves user experience by providing visual feedback on the presence of local metadata files.

* Improves metadata retrieval and UI updates

Refactors metadata retrieval logic for enhanced accuracy.
Specifically, improves local file annotation by adjusting search patterns to include package directories,
ensuring correct identification of files.

Updates the LWC UI to reorder columns for better usability.
Adds logic to unselect rows representing successfully retrieved metadata,
preventing redundant selections.

Fixes an issue where the panel was not being passed to the retrieve execution.

* Improves local file detection performance

Optimizes local file existence checks by indexing metadata
files. This significantly reduces file system operations.

It avoids redundant scans, and enhances speed when retrieving
metadata, especially when dealing with large projects.

* Improves metadata retrieval and conflict handling

Enhances the metadata retrieval process by adding conflict handling and more informative user feedback.

- Introduces a force overwrite option for metadata retrieval in case of source conflicts.
- Improves the user interface by displaying more specific error messages, including the name of the problematic metadata.
- Updates local file status indicators and adds download functionality directly within the webview.
- Improves the file search algorithm in the local file annotation process.

* Enables lazy-loading of metadata types

Introduces lazy-loading of metadata types in the metadata retriever to improve performance. It also adds `IsNameObsolete` to the SourceMember query to correctly identify deleted members.

The metadata types are now fetched when the org is selected, and when the org is changed, which reduces initial loading time and ensures that the list of available metadata types is up-to-date for the selected org.

* Enables metadata deletion from the UI

Allows users to delete metadata components directly from the UI.

This change introduces the ability to mark metadata items as "deleted" within the UI, triggering their removal from the local workspace upon retrieval. The action removes the corresponding files from the project directory to sync local files with the org.

Adds a confirmation message to the user upon successful deletion with a link to view and commit the changes.

* changelog

* Adds floating button for metadata retrieval

Improves the user experience by adding a floating button
to retrieve metadata when the main button is off-screen.

This ensures that users can easily retrieve metadata even
when they have scrolled down the page and the original
button is no longer visible.

* LWC UI PAnel is called everywhere, maybe it's better to send

* typo

* cspell

* [Mega-Linter] Apply linters fixes :)

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>