Persists metadata retrieval history (#293)
* Persists metadata retrieval history

Adds functionality to store a history of retrieved metadata in the `hardis-report/retrieve` directory.

This includes:
- Saving individual retrieval results as timestamped package.xml files.
- Maintaining a global package.xml file containing all retrieved metadata for cumulative tracking.

This provides users with a record of their metadata retrieval activities and allows for easier management of retrieved components.

* [Mega-Linter] Apply linters fixes

* Adds functionality to view retrieve history

Implements a button in the metadata retriever UI that opens
the retrieve folder in the VS Code explorer, allowing users
to easily access and review their metadata retrieve history.

* [Mega-Linter] Apply linters fixes

* lbl

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>