Enables saving deployment actions for PRs
Adds functionality to persist pre/post deployment commands defined for a pull request to a YAML file.

This allows users to define and save custom deployment actions (commands) that should be executed before or after a deployment for a specific pull request. The saved commands are stored in a `.sfdx-hardis.PR_NUMBER.yml` file.

Adds `uuid` dependency to generate unique IDs for deployment actions.