Adds missing options to select lists
Ensures that previously selected values for apex scripts and sfdmu workspaces, which might not be available in the current branch's options, are still displayed in the select lists. This prevents the user interface from losing track of the selected action parameter even if the selected value is not visible from current git branch.