Enables deployment actions in pipeline view (#299)
* Adds deployment actions to pipeline branches

Adds the ability to define pre/post deployment commands in pull requests via a configuration file.

These commands are then displayed in the pipeline UI, providing visibility into deployment-related actions.

This enhancement allows users to orchestrate and track deployment tasks more effectively.

* Improves action display and sorting

Refactors how actions are displayed to include a 'when' field indicating the deployment phase.

Updates sorting logic to prioritize actions based on the 'when' field (Pre-Deploy, Post-Deploy).

This ensures actions are displayed in the correct order relative to the deployment process.

* Adds deployment action modal and save functionality

Implements a modal to display and edit deployment actions, including action details and configuration options.

Introduces "saveDeploymentAction" message type to trigger the save logic in the extension.
Adds temporary save logic with information message to signal save functionality
is not yet implemented.

* fix

* Enables Apex script and SFDMU workspace selection

Enhances the deployment action configuration by allowing users to select Apex scripts and SFDMU workspaces from dropdown lists.

This feature provides a more user-friendly way to specify scripts and workspaces for deployment actions, improving the overall workflow.

The implementation includes:
- Fetching Apex scripts and SFDMU workspaces from the project.
- Displaying them as options in the deployment action configuration panel.

* label

* Enables saving deployment actions for PRs

Adds functionality to persist pre/post deployment commands defined for a pull request to a YAML file.

This allows users to define and save custom deployment actions (commands) that should be executed before or after a deployment for a specific pull request. The saved commands are stored in a `.sfdx-hardis.PR_NUMBER.yml` file.

Adds `uuid` dependency to generate unique IDs for deployment actions.

* fix

* Enhances pipeline view with current branch PR

Adds functionality to display and manage the current branch's pull request within the pipeline view.

This includes:
- Displaying the current PR in the main pipeline view.
- Allowing users to view related tickets and manage deployment actions for the current PR directly.
- Associating draft deployment action files with the current PR.
- Fetching PR info with tickets and deployment actions.
- UI enhancements for deployment action.

* Enhances PR draft actions file handling

Improves the handling of draft deployment actions files. It now associates the draft file with the current pull request or displays a message indicating that no pull request has been created yet.
This change also ensures that the correct file name is used when retrieving pre/post commands based on whether a PR exists.

Addresses an issue where draft files were not being correctly linked or handled when a PR didn't exist for the current branch.

* Adds missing options to select lists

Ensures that previously selected values for apex scripts and sfdmu workspaces, which might not be available in the current branch's options, are still displayed in the select lists. This prevents the user interface from losing track of the selected action parameter even if the selected value is not visible from current git branch.

* Enhances single PR view in pipeline

Improves the single PR view in the pipeline modal by adding a link to open the PR in the Git provider's web interface.

Also, fetches PR details to display a more informative title for the modal.

* beta

* changelog

* fixes

* add normalize-url

* [Mega-Linter] Apply linters fixes :)

* fix PR label

* Fixes Git/Jira icons display

Ensures Git and Jira icons are only rendered if their respective URLs are available. This prevents broken image icons from displaying when the connections are not configured.

* ml

* [Mega-Linter] Apply linters fixes :)

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>