Adds local package retrieve option
Adds the ability to retrieve metadata from a specific local sfdx package folder.

This change introduces a new option in the Metadata Retriever UI to select a local package directory from the sfdx-project.json file. This allows users to target a specific package within their project when retrieving metadata, instead of the entire project.

It also addresses an issue where the default package directory in sfdx-project.json could be inadvertently modified during the retrieval process. To prevent this, a guard is implemented to store the initial default package directory and restore it after the retrieval is complete, ensuring that the user's sfdx-project.json remains unchanged.

Fixes https://github.com/hardisgroupcom/sfdx-hardis/issues/1599